requirements.txt
selenium==4.23.1
allure-pytest==2.14.3
pytest==6.2.5
requests==2.32.3
faker==33.1.0